adding public pages

+++ code_py2.7/frontend/serve.py
@@ -76,11 +76,9 @@ try:
                         pass
 
             print '[*] %s wants to create an account' % client_addr[0]
-            client.send(open(html_engine.display_information(client_addr[0], user_agent),'rb').read())
+        elif 'GET /favicon.ico HTTP/1.1' in request.split('\r\n'):
+            client.send(open(html_engine.display_information(client_addr[0], user_agent), 'rb').read())
             os.remove('info.html')
-            time.sleep(0.4)   # These requests seem to come in multiples from the browser?
-        else:
-            print request
         # Login attempts
         if len(request.split('username=')) > 1:
             uname = request.split('username=')[1].split('&')[0]
